testing only fp16 and int8

+++ quantization/sanity_check.py
@@ -7,7 +7,7 @@ bnb_config = BitsAndBytesConfig(
     load_in_8bit=True)
 
 #model_names = ["fp16", "int8", "gptq", "awq"]
-model_names = ["fp16", "int8"]
+model_names = ["int8", "fp16"]
 for model_name  in model_names:
     print(f"Loading {model_name }...")
     if model_name == "fp16":
